Updating user-defined voice processing.
All user-defined voices are still based off of default voice Paul.
Initially, this was done by using the [:dv] command to set all of Paul's
default voice parameters. This could cause problems if the default
values are incorrect (since the language spec was not based on source
code information and could be incorrect) or if the text-to-speech engine
does not support every voice parameter mentioned in the language
specification.

For the latter case, the text-to-speech engine is still expected to
support every voice parameter in the opedec language specification.
However, it's a nice feature to still be able to use a
partially-compliant engine when calling user-defined voices.